feat: add `Source` get method
